Added extra tag called 'ldap' with isc-dhcp-server-ldap package installed
Also added additional build and run scripts for that variant.

This way of doing it duplicates a lot of code, but it does work. If anyone
has an idea of how to only use a single Dockerfile with an environment
variable or something like that to avoid the code duplication then I'd
very much like a pull request that solves this issue. Same goes for the
build and run scripts.